services/faceid/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
insightface==0.7.3
onnxruntime==1.19.2
numpy==1.26.4
Pillow==10.4.0

